refactor: add new util file
